+++ NEWS.md
@@ -39,7 +39,7 @@ output: html_document
    When `TRUE`, a warning is issued if argument `a` is missing entries on `margin`
    that are present in `mat`.
 * Many new tests for new features.
-    - Now at 1036 tests, all passing.
+    - Now at 1039 tests, all passing.
     - Test coverage remains at 100 %.